Move ExploreRootState to explore type file
